agent-codex-gpt51/pyproject.toml
[project]
name = "agent-codex-gpt51"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.8",
    "langgraph>=1.0.4",
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langsmith>=0.4.48",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "ruff>=0.14.6",
]
